Create challengeFive.js
creates a game where a divorced man has to collect divorce papers, functions are used to draw both him and the divorce papers, as well as the background which is meant to represent the office he's in. Functions are used to calculate collision for the sprites, so a new divorce paper will appear  in a random spot on the top of the screen every time he collects one. In this version, he frowns when he collects a divorce paper.